test(payment): 통합 테스트와 단위 테스트 분리
